This is the cleaned up version of the Application. I've added all the methods neccessary and checked that the GUI launches. DAVE: Please fix the issue that fetches the text from the fields. Its showing an error on line 57 in CONTROLLER.JAVA file
UPDATES:
Removed SongList class b/c it was making the listview difficult. So i implemented an ObservableList<Songs> instead to store the objects of type song.
Created XML file that will be used to save and retrieve data to and from ObservableList<Songs>
Copied over code from Daves GUI(fxml code and Main contents)

Tested that gui launches but the add button does not work

Things to do:
Make the three buttons work(ADD/EDIT/DELETE)
Populate the FULL details of data in the labels when a song is selected.
BTW only song name and artist should be displayed on the LIST VIEW.!!!!!!!!!!!!!!!!!
Test for errors